UX changes on the android app
This commit adds in UX changes. The app should now show an alert dialog
when the user successfully subscribes to a tag. The app will also show
an alert dialog when a city is found.

This commit also modifies the app to use the credentials in the xml file
rather than inside a java file.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>